tools/syz-diff: move the logic to pkg/manager
